+++ client/src/App.js
@@ -1,5 +1,7 @@
 import './App.css';
 import {Switch, Route, Link} from 'react-router-dom';
+import User from './components/providers/User';
+import {GoogleSignInButton} from './components/GoogleSignInButton';
 
 // components
 import RegisterServiceForm from './components/RegisterServiceForm';
@@ -13,54 +15,57 @@ function App() {
 
   return (
     <div className="App">
-      {!loggedIn ? (
-        <h1> Landing page</h1>
-      ) : (
-        <div>
-          <Header className="header">
-            <div className="logo" />
-            <Menu theme="dark" mode="horizontal" defaultSelectedKeys={['3']}>
-              <Menu.Item key="1">
-                <Link to="/">Home</Link>
-              </Menu.Item>
-              <Menu.Item key="2">
-                <Link to="/register-service">Register Service</Link>
-              </Menu.Item>
-              <Menu.Item key="3">
-                <Link to="/dashboard"> Dashboard </Link>
-              </Menu.Item>
-            </Menu>
-          </Header>
+      <User.Provider>
+        {!loggedIn ? (
+          <h1> Landing page</h1>
+        ) : (
+          <div>
+            <Header className="header">
+              <div className="logo" />
+              <Menu theme="dark" mode="horizontal" defaultSelectedKeys={['3']}>
+                <Menu.Item key="1">
+                  <Link to="/">Home</Link>
+                </Menu.Item>
+                <Menu.Item key="2">
+                  <Link to="/register-service">Register Service</Link>
+                </Menu.Item>
+                <Menu.Item key="3">
+                  <Link to="/dashboard"> Dashboard </Link>
+                </Menu.Item>
+              </Menu>
+            </Header>
 
-          <Layout>
-            <SideMenu />
             <Layout>
-              <Layout style={{padding: '0 24px 24px'}}>
-                <Content
-                  className="site-layout-background"
-                  style={{
-                    padding: 24,
-                    marginTop: 24,
-                    minHeight: 280,
-                  }}>
-                  <Switch>
-                    <Route path="/dashboard">
-                      <Projects />
-                    </Route>
-                    <Route path="/register-service">
-                      <RegisterServiceForm />
-                    </Route>
+              <SideMenu />
+              <Layout>
+                <Layout style={{padding: '0 24px 24px'}}>
+                  <Content
+                    className="site-layout-background"
+                    style={{
+                      padding: 24,
+                      marginTop: 24,
+                      minHeight: 280,
+                    }}>
+                    <Switch>
+                      <Route path="/dashboard">
+                        <Projects />
+                      </Route>
+                      <Route path="/register-service">
+                        <RegisterServiceForm />
+                      </Route>
 
-                    <Route path="/">
-                      <h1>Mosaic</h1>
-                    </Route>
-                  </Switch>
-                </Content>
+                      <Route path="/">
+                        <h1>Mosaic</h1>
+                      </Route>
+                    </Switch>
+                  </Content>
+                </Layout>
               </Layout>
             </Layout>
-          </Layout>
-        </div>
-      )}
+            <GoogleSignInButton />
+          </div>
+        )}
+      </User.Provider>
     </div>
   );
 }