+++ server/models/ServiceModel.js
@@ -1,11 +1,13 @@
-const mongoose = require('mongoose');
-const Schema = mongoose.Schema;
+import mongoose from 'mongoose';
+import {options} from './helpers';
 
-const options = {
-  timestamps: true,
-  createdAt: 'created_at',
-  updatedAt: 'updated_at',
-};
+const Schema = mongoose.Schema;
+const {
+  Types: {
+    ObjectId,
+    String,
+  }
+} = Schema;
 
 const ServiceSchema = new Schema(
   {
@@ -31,17 +33,19 @@ const ServiceSchema = new Schema(
       required: true,
     },
     dependency: {
-      type: Array,
+      type: [ObjectId],
+      default: [],
     },
     events: {
-      type: Array,
+      type: [ObjectId],
+      default: [],
     },
     description: {
       type: String,
-      required: true,
+      default: "",
     },
     projectId: {
-      type: String,
+      type: ObjectId,
       required: true,
     },
   },
@@ -49,4 +53,4 @@ const ServiceSchema = new Schema(
 );
 
 const Service = mongoose.model('services', ServiceSchema);
-module.exports = Service;
+export default {Service};