+++ server/changelog.md
@@ -1,2 +1,8 @@
 <!-- add PR # and description -->
-
+### May 22
+- [] Update config so that the backend repo, you can now use ES6 module exports/imports instead of commonJS modules
+- [] Replace JS method of hiding keys & credentials with industry standard environment variables (env file & dotenv module)
+- [] Add SIgn In With Google using passportJS, express-session, connect-mongo, etc.
+- [] Fix issues with models (naming inconsistencies, mongo options, mongo types, exports/imports)
+- [] Refactor & restructure backend repo to have proper directory hierarchy & code splitting
+- [] Add proxy to be able to take in requests from frontend react server